Commmit 1 20201013

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -7,8 +7,12 @@ import org.hibernate.Session;
 import org.hibernate.Transaction;
 
 import java.util.List;
+import java.util.logging.Logger;
 
 public class UserDaoHibernateImpl implements UserDao {
+    private static Session session;
+    public final static Logger logger = Logger.getLogger(String.valueOf(UserDaoHibernateImpl.class));
+
     public UserDaoHibernateImpl() {
 
     }
@@ -16,7 +20,26 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public void createUsersTable() {
+        try {
+            session = Util.getSessionFactory().openSession();
+            session.beginTransaction();
 
+            Query queryObj = session.createQuery(CREATE_TABLE_QUERY);
+            queryObj.executeUpdate();
+
+            session.getTransaction().commit();
+            logger.info("\nSuccessfully Deleted All Records From The Database Table!\n");
+        } catch (Exception sqlException) {
+            if (null != session.getTransaction()) {
+                logger.info("\n.......Transaction Is Being Rolled Back.......\n");
+                session.getTransaction().rollback();
+            }
+            sqlException.printStackTrace();
+        } finally {
+            if (session != null) {
+                session.close();
+            }
+        }
     }
 
     @Override
@@ -27,26 +50,28 @@ public class UserDaoHibernateImpl implements UserDao {
     @Override
     public void saveUser(String name, String lastName, byte age) {
         User user = new User(name, lastName, age);
-        Session session = Util.getSessionFactory().openSession();
-        Transaction tx1 = session.beginTransaction();
+        session = Util.getSessionFactory().openSession();
+        session.beginTransaction();
         session.save(user);
-        tx1.commit();
+        session.getTransaction().commit();
         session.close();
     }
 
     @Override
     public void removeUserById(long id) {
         User user = findUserById(id);
-        Session session = Util.getSessionFactory().openSession();
-        Transaction tx1 = session.beginTransaction();
-        session.delete(user);
-        tx1.commit();
-        session.close();
+        if (user != null) {
+            session = Util.getSessionFactory().openSession();
+            session.beginTransaction();
+            session.delete(user);
+            session.getTransaction().commit();
+            session.close();
+        }
     }
 
     @Override
     public List<User> getAllUsers() {
-        List<User> users = (List<User>)  Util.getSessionFactory().openSession().createQuery("From User").list();
+        List<User> users = (List<User>) Util.getSessionFactory().openSession().createQuery("From User").list();
         return users;
     }
 
@@ -54,31 +79,45 @@ public class UserDaoHibernateImpl implements UserDao {
     public void cleanUsersTable() {
 //        https://examples.javacodegeeks.com/enterprise-java/hibernate/hibernate-crud-operations-tutorial
         try {
-            // Getting Session Object From SessionFactory
-            Session session = Util.getSessionFactory().openSession();
-            // Getting Transaction Object From Session Object
+            session = Util.getSessionFactory().openSession();
             session.beginTransaction();
 
             Query queryObj = session.createQuery("DELETE FROM Student");
             queryObj.executeUpdate();
 
-            // Committing The Transactions To The Database
             session.getTransaction().commit();
-//            logger.info("\nSuccessfully Deleted All Records From The Database Table!\n");
-        } catch(Exception sqlException) {
-            if(null != sessionObj.getTransaction()) {
+            logger.info("\nSuccessfully Deleted All Records From The Database Table!\n");
+        } catch (Exception sqlException) {
+            if (null != session.getTransaction()) {
                 logger.info("\n.......Transaction Is Being Rolled Back.......\n");
-                sessionObj.getTransaction().rollback();
+                session.getTransaction().rollback();
             }
             sqlException.printStackTrace();
         } finally {
-            if(sessionObj != null) {
-                sessionObj.close();
+            if (session != null) {
+                session.close();
             }
         }
     }
 
     public User findUserById(long id) {
-        return Util.getSessionFactory().openSession().get(User.class, id);
+        User user = null;
+        try {
+            session = Util.getSessionFactory().openSession();
+            session.beginTransaction();
+
+            user = (User) session.get(User.class, id);
+        } catch (Exception sqlException) {
+            if (null != session.getTransaction()) {
+                logger.info("\n.......Transaction Is Being Rolled Back.......\n");
+                session.getTransaction().rollback();
+            }
+            sqlException.printStackTrace();
+        }
+        return user;
+    }
+
+    public void closeConnection() {
+
     }
 }
